Move remainder of MCPL-related code to mcpl_interface.h/cpp
