+++ CMakeLists.txt
@@ -164,6 +164,7 @@ endif()
 #===============================================================================
 if (OPENMC_USE_MCPL)
   find_package(MCPL REQUIRED)
+  message(STATUS "Found MCPL: ${MCPL_DIR} (found version \"${MCPL_VERSION}\")")
 endif()
 
 #===============================================================================
